Complete refactor update: use step file, add information for user.
Update process has majors drawbacks:

 - we skip a lot tasks because we loop over the same update step task
   file, changing the step variable, wasting time and clobbering logs.

 - it's hard to tell in which part of the process the update failed
   for the user (and even for devs)

 - update run tasks are from an include, preventing any possibility to
   be able to use the start-at-tasks ansible feature.

To solve the skipped tasks this, we now loop over the stepX file
generated for update and post update.

To solve the lack of clear feedback to the user we enclose each import
with "Start/Finish" debug message whose task name has the "human" name
of the current task being done during update.  In case of an error
we ensure that we clearly state in which part of the update we
failed.  That will ease communication when issue arise.

To be able to use start-at-tasks in the future we ensure that we only
use import_task.

Eventually, to avoid to cluter deploy.j2 with too many code repetition,
we use a macro to generate the necessary ansible code.

WIP: NOT sure we should do that as we the package are not yet updated
at that point.  There is also a slight change in the ordering of the
tasks where host prep task is run much earlier in the process.

Depends-On: https://review.opendev.org/740247
Change-Id: Ib32791d72410766a0b2e2330713120d15dc35f57